task-08/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
PySide6==6.5.2
PySide6-Addons==6.5.2
PySide6-Essentials==6.5.2
requests==2.31.0
shiboken6==6.5.2
urllib3==2.0.4
Pillow==8.3.2